+++ src/main/java/pa/model/OrganizationRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrganizationRepository extends CrudRepository<Organization, Integer> {
+public interface OrganizationRepository extends CrudRepository<Organization, Long> {
 }
