+++ src/main/java/pa/model/RegionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RegionRepository extends CrudRepository<Region, Integer> {
+public interface RegionRepository extends CrudRepository<Region, Long> {
 	@Override
 	List<Region> findAll();
 }
